+++ services/shape_detection/android/java/src/org/chromium/shape_detection/InterfaceRegistrar.java
@@ -15,12 +15,12 @@ import org.chromium.shape_detection.mojom.TextDetection;
 
 @JNINamespace("shape_detection")
 class InterfaceRegistrar {
-    static MessagePipeHandle messagePipeHandleFromNative(int nativeHandle) {
+    static MessagePipeHandle messagePipeHandleFromNative(long nativeHandle) {
         return CoreImpl.getInstance().acquireNativeHandle(nativeHandle).toMessagePipeHandle();
     }
 
     @CalledByNative
-    static void bindBarcodeDetectionProvider(int nativeHandle) {
+    static void bindBarcodeDetectionProvider(long nativeHandle) {
         // Immediately wrap |nativeHandle| as it cannot be allowed to leak.
         MessagePipeHandle handle = messagePipeHandleFromNative(nativeHandle);
 
@@ -34,13 +34,13 @@ class InterfaceRegistrar {
     }
 
     @CalledByNative
-    static void bindFaceDetectionProvider(int nativeHandle) {
+    static void bindFaceDetectionProvider(long nativeHandle) {
         FaceDetectionProvider.MANAGER.bind(
                 new FaceDetectionProviderImpl(), messagePipeHandleFromNative(nativeHandle));
     }
 
     @CalledByNative
-    static void bindTextDetection(int nativeHandle) {
+    static void bindTextDetection(long nativeHandle) {
         // Immediately wrap |nativeHandle| as it cannot be allowed to leak.
         MessagePipeHandle handle = messagePipeHandleFromNative(nativeHandle);
